requirements.txt
Django==4.2.7
GitPython==3.1.40
python-decouple==3.8
Pillow==10.1.0
WeasyPrint==60.2
reportlab==4.0.7
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-ses==3.5.0
boto3==1.34.0
